Correctly identifying \r as whitespace.
Using | (bitwise OR) instead of || (logical OR) would fail to
detect \r as whitespace.

This is a whitespace fix to `TinyXML::processChar()`. The | (bitwise OR)
operator has a higher precedence the || (logical OR). The result is that
`<any character> | '\r'` is always true, and the order precedence bound
true with the prior `ch == '\n'` effectively making it `ch == '\n' | true`,
so `ch | '\r'` was a no-op. What this change does mean is that previously
\r was not whitespace, but now it is.